agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[tavus]~=1.3",
    "python-dotenv>=1.2.1",
]
